ur kidding right

+++ src/client/cl_tc_vis.cpp
@@ -333,7 +333,7 @@ static void add_triggers(void) {
 					Com_Printf("tc_vis_init: adding trigger brush, i: %d \n", i);
 				}
 				if (com_developer->integer > 2) {
-					Com_Printf("tc_vis_init: color: %d %d %d %d, ptr: %d \n", (int)trigger_color.b[0], (int)trigger_color.b[1], (int)trigger_color.b[2], (int)trigger_color.b[3],(int)& trigger_color);
+					Com_Printf("tc_vis_init: color: %d %d %d %d, ptr: %d \n", (int)trigger_color.b[0], (int)trigger_color.b[1], (int)trigger_color.b[2], (int)trigger_color.b[3],(int)(size_t)&trigger_color);
 				}
 				gen_visible_brush(cm.leafbrushes[leaf->firstLeafBrush + i], origin, TRIGGER_BRUSH, &trigger_color);
 			}
